리뷰 생성/삭제 시 관련 Entity 업데이트 로직 추가

+++ matgpt/src/main/java/com/ktc/matgpt/review/dto/ReviewRequest.java
@@ -76,14 +76,14 @@ public class ReviewRequest {
     public static class CreateDTO {
         private List<ImageDTO> reviewImages;
         private String content;
-        private double rating;
+        private int rating;
         @Min(1)
         private int peopleCount;
         @Min(0)
         private int totalPrice;
 
         public CreateDTO(List<ImageDTO> reviewImages, String content,
-                         double rating, int peopleCount, int totalPrice) {
+                         int rating, int peopleCount, int totalPrice) {
             this.reviewImages = reviewImages;
             this.content = content;
             this.rating = rating;
